Subscribe to routting events before they are triggered
