+++ src/app/modules/gb-main-module/gb-main.routing.ts
@@ -10,7 +10,7 @@ import {ModuleWithProviders} from '@angular/core';
 import {Routes, RouterModule} from '@angular/router';
 
 // Route Configuration
-export const routes: Routes = [
+const routes: Routes = [
   {
     path: '',
     redirectTo: '/cohort-selection',
